+++ lib/widgets/deleted_list_item.dart
@@ -11,7 +11,7 @@ class DeletedListItem extends StatelessWidget {
   final String fullName;
   final String phone;
   final int tour;
-  final int poids;
+  final double poids;
   final int nombrePalettes;
   final int nombreSac;
   final DateTime createdAt;
